+++ CHANGELOG.md
@@ -70,4 +70,4 @@ IMPROVEMENTS:
 
 ## 1.0.0 (November 01, 2017)
 
-* Initial release of the CloudScale.ch provider
+* Initial release of the cloudscale.ch provider